[fuzzer] Extract common driver code
So that h3 fuzzer can reuse it later.
