+++ CMakeLists.txt
@@ -863,8 +863,8 @@ IF (BUILD_FUZZER)
     IF(NOT CMAKE_CXX_COMPILER_ID STREQUAL "Clang")
         MESSAGE(FATAL_ERROR "The fuzzer needs clang as a compiler")
     ENDIF()
-    ADD_EXECUTABLE(h2o-fuzzer-http1 fuzz/driver.cc)
-    ADD_EXECUTABLE(h2o-fuzzer-http2 fuzz/driver.cc)
+    ADD_EXECUTABLE(h2o-fuzzer-http1 fuzz/driver.cc fuzz/driver_common.cc)
+    ADD_EXECUTABLE(h2o-fuzzer-http2 fuzz/driver.cc fuzz/driver_common.cc)
     ADD_EXECUTABLE(h2o-fuzzer-http3 fuzz/driver_h3.cc fuzz/quicly_mock.c ${FUZZED_SOURCE_FILES})
     ADD_EXECUTABLE(h2o-fuzzer-url fuzz/driver_url.cc)
     ADD_EXECUTABLE(h3-header-generator fuzz/h3_header_generator.c)
